Baselines - One to Rule Them All
It seems heavyweight to have a separate baseline for GT and adds a lot of error-prone duplication; each change to project structure requires carefully syncing changes to each baseline. The baseline could probably be simplified even more if this refactoring is propogated down into the dependencies, which seem to use the same multi-baseline pattern.